add default attrs

+++ risenumbertextview/src/main/java/com/chrischeng/risenumbertextview/NumberType.java
@@ -1,6 +1,6 @@
 package com.chrischeng.risenumbertextview;
 
 public enum NumberType {
-    INTEGER,
-    FLOAT;
+    DECIMAL,
+    INTEGER;
 }
